crawler/requirements.txt
mysql
scrapy
